Updates for recent DZ files
Additional fields in chunks and header are now getting used.  Turns out
what appears to be an Android version field is longer than had been
thought.  A new portion of the header is now sometimes 256.

Open image files as overwrite.  This is useful to update an existing
device image that was grabbed by other means.

Make the header consistency check report both computed and expected sizes

to make analysis easier in future.