Adjust out of order warning
The chunks are roughly in the order they get written to a device.  A key
issue is areas that are written in second pass appear after chunks that
are written towards the end.  Make sure this warning doesn't appear for
non-zero order chunks (unless they're out of order with respect to chunks
in their own pass).  Move to always sorting since we're more interested
in the order chunks actually appear in.